+++ src/backend_cairo/mod.rs
@@ -37,6 +37,7 @@ mod fill;
 mod filter;
 mod gradient;
 mod image;
+mod marker;
 mod mask;
 mod path;
 mod pattern;
@@ -264,7 +265,7 @@ fn render_node(
             Some(render_group(node, opt, layers, cr))
         }
         usvg::NodeKind::Path(ref path) => {
-            Some(path::draw(&node.tree(), path, opt, cr))
+            Some(path::draw(&node.tree(), path, opt, layers, cr))
         }
         usvg::NodeKind::Text(ref text) => {
             Some(text::draw(&node.tree(), text, opt, cr))