+++ src/backend_cairo/path.rs
@@ -10,6 +10,7 @@ use super::prelude::*;
 use super::{
     fill,
     stroke,
+    marker,
 };
 
 
@@ -17,6 +18,7 @@ pub fn draw(
     tree: &usvg::Tree,
     path: &usvg::Path,
     opt: &Options,
+    layers: &mut CairoLayers,
     cr: &cairo::Context,
 ) -> Rect {
     let mut is_square_cap = false;
@@ -41,6 +43,8 @@ pub fn draw(
     } else {
         cr.fill();
     }
+
+    marker::apply(tree, path, opt, layers, cr);
 
     bbox
 }