+++ src/backend_qt/pattern.rs
@@ -23,6 +23,7 @@ pub fn apply(
         pattern.rect
     };
 
+    // TODO: wrong
     let global_ts = usvg::Transform::from_native(&global_ts);
     let (sx, sy) = global_ts.get_scale();
     // Only integer scaling is allowed.