+++ src/backend_utils/mod.rs
@@ -4,6 +4,7 @@
 
 pub mod filter;
 pub mod image;
+pub mod marker;
 pub mod mask;
 pub mod text;